requirements.txt
pytest==7.1.3
python-dotenv==0.21.0
requests==2.28.1
